FIN-124: UI) 메뉴 - 약관 동의 디자인 변경 사항 요청
